ACTIN-1467: Consolidate gene driver and curated file reading
